refactor(onboarding): rename components, move overlays into components
