+++ src/scripts/components/main/onboarding-tooltip/onboarding-tooltip.styl
@@ -1,5 +1,8 @@
 @require '../../../../variables.styl'
 
+.onboardingOverlay
+  background: transparent
+
 .onboarding
   position: absolute
   display: flex
