+++ src/scripts/components/main/welcome-screen/welcome-screen.styl
@@ -1,5 +1,8 @@
 @require '../../../../variables.styl'
 
+.welcomeOverlay
+  background: linear-gradient(90deg, rgba(16, 22, 26, 0.85) 0%, rgba(16, 22, 26, 0.81) 48%, rgba(16, 22, 26, 0.79) 50%, rgba(16, 22, 26, 0.75) 52%, rgba(16, 22, 26, 0.2) 60%, rgba(16, 22, 26, 0.1) 70%, rgba(16, 22, 26, 0) 100%)
+
 .welcomeScreen
   display: flex
   flex-direction: row
@@ -51,6 +54,9 @@
   background: rgba($main, 0.2)
 
 @media screen and (max-width: 480px)
+  .welcomeOverlay
+    background: linear-gradient(rgba(16, 22, 26, 0.85) 0%, rgba(16, 22, 26, 0.81) 48%, rgba(16, 22, 26, 0.79) 50%, rgba(16, 22, 26, 0.75) 52%, rgba(16, 22, 26, 0.2) 60%, rgba(16, 22, 26, 0.1) 70%, rgba(16, 22, 26, 0) 100%)
+
   .welcomeScreen
     width: 100%
     height: 100%
