+++ src/scripts/components/main/welcome-screen/welcome-screen.tsx
@@ -5,6 +5,7 @@ import {useDispatch} from 'react-redux';
 
 import setWelcomeScreenAction from '../../../actions/set-welcome-screen';
 import Button from '../button/button';
+import Overlay from '../overlay/overlay';
 
 import styles from './welcome-screen.styl';
 
@@ -17,42 +18,44 @@ const WelcomeScreen: FunctionComponent<Props> = ({onStartOnboarding}) => {
   const dispatch = useDispatch();
 
   return (
-    <div className={styles.welcomeScreen}>
-      <div className={styles.content}>
-        <ReactMarkdown
-          source={intl.formatMessage({id: 'welcomeContent'})}
-          linkTarget="_blank"
-          allowedTypes={[
-            'heading',
-            'text',
-            'paragraph',
-            'break',
-            'strong',
-            'emphasis',
-            'image',
-            'imageReference',
-            'list',
-            'listItem',
-            'link'
-          ]}
-        />
-        <div className={styles.tourButtons}>
-          <Button
-            className={styles.secondaryTourButton}
-            label="cancelTour"
-            onClick={() => dispatch(setWelcomeScreenAction(false))}
+    <Overlay className={styles.welcomeOverlay} showCloseButton={false}>
+      <div className={styles.welcomeScreen}>
+        <div className={styles.content}>
+          <ReactMarkdown
+            source={intl.formatMessage({id: 'welcomeContent'})}
+            linkTarget="_blank"
+            allowedTypes={[
+              'heading',
+              'text',
+              'paragraph',
+              'break',
+              'strong',
+              'emphasis',
+              'image',
+              'imageReference',
+              'list',
+              'listItem',
+              'link'
+            ]}
           />
-          <Button
-            className={styles.primaryTourButton}
-            label="startTour"
-            onClick={() => {
-              dispatch(setWelcomeScreenAction(false));
-              onStartOnboarding();
-            }}
-          />
+          <div className={styles.tourButtons}>
+            <Button
+              className={styles.secondaryTourButton}
+              label="cancelTour"
+              onClick={() => dispatch(setWelcomeScreenAction(false))}
+            />
+            <Button
+              className={styles.primaryTourButton}
+              label="startTour"
+              onClick={() => {
+                dispatch(setWelcomeScreenAction(false));
+                onStartOnboarding();
+              }}
+            />
+          </div>
         </div>
       </div>
-    </div>
+    </Overlay>
   );
 };
