+++ src/scripts/components/main/welcome/welcome.tsx
@@ -1,12 +1,9 @@
 import React, {FunctionComponent, useState} from 'react';
 import {useSelector} from 'react-redux';
+
 import {welcomeScreenSelector} from '../../../selectors/welcome-screen';
-
-import OnboardingTooltip from '../onboarding/onboarding';
-import Overlay from '../overlay/overlay';
 import WelcomeScreen from '../welcome-screen/welcome-screen';
-
-import styles from './welcome.styl';
+import OnboardingTooltip from '../onboarding-tooltip/onboarding-tooltip';
 
 const Onboarding: FunctionComponent = () => {
   const [onboardingStep, setOnboardingStep] = useState<number | null>(null);
@@ -15,19 +12,15 @@ const Onboarding: FunctionComponent = () => {
   return (
     <React.Fragment>
       {showWelcomeScreen && (
-        <Overlay className={styles.welcomeOverlay} showCloseButton={false}>
-          <WelcomeScreen onStartOnboarding={() => setOnboardingStep(1)} />
-        </Overlay>
+        <WelcomeScreen onStartOnboarding={() => setOnboardingStep(1)} />
       )}
 
       {onboardingStep && (
-        <Overlay className={styles.onboardingOverlay} showCloseButton={false}>
-          <OnboardingTooltip
-            step={onboardingStep}
-            onPageChange={(step: number) => setOnboardingStep(step)}
-            onClose={() => setOnboardingStep(null)}
-          />
-        </Overlay>
+        <OnboardingTooltip
+          step={onboardingStep}
+          onPageChange={(step: number) => setOnboardingStep(step)}
+          onClose={() => setOnboardingStep(null)}
+        />
       )}
     </React.Fragment>
   );
